Move MMT APIs to one folder and improve the mmirs API.
