+++ skyportal/facility_apis/__init__.py
@@ -23,8 +23,8 @@ from .tess import TESSAPI
 from .trt import TRTAPI
 from .winter import WINTERAPI
 from .ztf import ZTFAPI, ZTFMMAAPI
-from .binospec import BINOSPECAPI
-from .mmirs import MMIRSAPI
+from .mmt.binospec import BINOSPECAPI
+from .mmt.mmirs import MMIRSAPI
 
 
 # APIs, listed in implementation order, to keep matching enum in db stable
